+++ ChangeLog.md
@@ -3,6 +3,12 @@
 ## REVISION HISTORY
 
 ---
+- 2017-08-30 Version 5.6.9
+  - Set the default timer to gettimeofday()
+  - Move the array initialization from start() to setProperties(), which maybe redundant but safer.
+  - Changed output info from DEBUG options
+
+---
 - 2017-08-29 Version 5.6.8
   - revert missing PMlib-user-manual.pdf